Renaming the main plugin file
